refactor(dashboard): extract subagent helpers to subagent.rs (#1266)
routes.rs: 2176 -> 2114
subagent.rs: NEW 66

Moves file_metrics, count_json_records, disk_usage_pct, subagent_sessions.
Updates metrics.rs to import file_metrics/count_json_records from new module.
