+++ src/operator_commands_dashboard/routes.rs
@@ -16,6 +16,7 @@ use super::distributed::{distributed, strip_ansi_codes, vacate_vm};
 use super::tmux::{azlin_tmux_sessions, ws_tmux_attach_handler};
 use super::chat::ws_chat_handler;
 use super::logs::read_tail;
+use super::subagent::{disk_usage_pct, file_metrics, count_json_records, subagent_sessions};
 use super::agent_log::{WS_AGENT_LOG_ROUTE, ws_agent_log_handler};
 use super::metrics::{memory_metrics, ooda_thinking};
 use super::registry::{agent_graph, build_lock_force_release, build_lock_status, registry_deregister, registry_list, registry_register, registry_reap};
@@ -241,69 +242,6 @@ pub(crate) fn resolve_state_root() -> std::path::PathBuf {
             let home = std::env::var("HOME").unwrap_or_else(|_| "/home/azureuser".to_string());
             std::path::PathBuf::from(home).join(".simard")
         })
-}
-
-
-pub(crate) fn file_metrics(path: &std::path::Path) -> (u64, Option<String>) {
-    match std::fs::metadata(path) {
-        Ok(m) => {
-            let size = m.len();
-            let modified = m.modified().ok().map(|t| {
-                let dt: chrono::DateTime<chrono::Utc> = t.into();
-                dt.to_rfc3339()
-            });
-            (size, modified)
-        }
-        Err(_) => (0, None),
-    }
-}
-
-pub(crate) fn count_json_records(path: &std::path::Path) -> u64 {
-    let Ok(content) = std::fs::read_to_string(path) else {
-        return 0;
-    };
-    match serde_json::from_str::<Value>(&content) {
-        Ok(Value::Array(arr)) => arr.len() as u64,
-        Ok(Value::Object(map)) => map.len() as u64,
-        _ => 0,
-    }
-}
-
-async fn disk_usage_pct() -> Option<u8> {
-    let output = tokio::process::Command::new("df")
-        .args(["--output=pcent", "/home"])
-        .output()
-        .await
-        .ok()?;
-    let text = String::from_utf8_lossy(&output.stdout);
-    let line = text.lines().nth(1)?;
-    line.trim().trim_end_matches('%').parse().ok()
-}
-
-
-/// WS-2: list live and recently-ended subagent tmux sessions.
-///
-/// Returns `{ live: [...], recently_ended: [...] }` sorted by `created_at`
-/// descending. The dashboard polls this every 5s to populate the Subagent
-/// Sessions card and to drive Attach deep-links in the Recent Actions feed.
-async fn subagent_sessions() -> Json<Value> {
-    let reg = crate::subagent_sessions::load();
-    let mut live: Vec<&crate::subagent_sessions::SubagentSession> = reg
-        .sessions
-        .iter()
-        .filter(|s| s.ended_at.is_none())
-        .collect();
-    let mut ended: Vec<&crate::subagent_sessions::SubagentSession> = reg
-        .sessions
-        .iter()
-        .filter(|s| s.ended_at.is_some())
-        .collect();
-    live.sort_by_key(|s| std::cmp::Reverse(s.created_at));
-    ended.sort_by_key(|s| std::cmp::Reverse(s.created_at));
-    Json(json!({
-        "live": live,
-        "recently_ended": ended,
-    }))
 }
 
 pub(crate) const INDEX_HTML: &str = r#"<!DOCTYPE html>
